+++ tests/integration/PaymentMethodTest.php
@@ -1170,6 +1170,7 @@ class PaymentMethodTest extends Setup
 
     public function testUpdate_acceptAccountTypeDebit()
     {
+        $this->markTestSkipped("Pending as flaky, until CI is more stable");
         $customer = Braintree\Customer::createNoValidate();
         $creditCardResult = Braintree\CreditCard::create([
             'cardholderName' => 'Original Holder',